cambiado a ft_echo
